requirements.txt
spotipy>=2.18.0
pandas>=1.1
numpy>=1.1
tensorflow-gpu>=2.4.0
scikit-learn>=0.24.2
Keras>=2.4.3
setuptools>=57.0.0